// types.d.ts
export class GPUBuffer {
    label: string | null;
    mapAsync( mode: GPUMapModeFlags, offset?: number, size?: number ): Promise<undefined>;
    getMappedRange(offset?: number, size?: number): ArrayBuffer;
    unmap(): undefined;
    destroy(): undefined;
}

export type GPUBindGroupEntry = {
    binding: number,
    resource: {
      buffer: GPUBuffer,
    }
}

export type GPUBindGroupDescriptor = {
    layout: GPUBindGroupLayout,
    entries: GPUBindGroupEntry[],
    label?: string
}

export type JobSize = {
    x: number,
    y?: number,
    z?: number
}

export type BufferBind = {
    buffer1: GPUBuffer,
    buffer2: GPUBuffer,
    size: number,
    offsets: [number, number]
}